+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.2.0 (July 22, 2021)
+
+ENHANCEMENTS:
+* Added `tfsdk.NewProtocol6Server` to return a `tfprotov6.ProviderServer` implementation for testing and muxing purposes. ([#72](https://github.com/hashicorp/terraform-plugin-framework/issues/72))
+* Added support for MapNestedAttributes. ([#79](https://github.com/hashicorp/terraform-plugin-framework/issues/79))
+* Responses now default to returning the current state, meaning state will only change when provider developers actively change it. Previously, an empty state value would be returned, which caused problems. ([#74](https://github.com/hashicorp/terraform-plugin-framework/issues/74))
+
 # 0.1.0 (June 24, 2021)
 
 FEATURES: